Encapsulated mapbox-gl into MapContainer component to be able to mock it
And both MapContainer and OrderForm into OrderManager component to avoid
re-rendering of the Header on each mouse click.
