+++ loft-taxi/src/components/MapView/MapContainer/style.scss
@@ -1,0 +1,6 @@
+@import 'shared/style';
+
+.loft-taxi-map-container {
+  width: 100%;
+  height: 100vh - $header-height;
+}
