+++ loft-taxi/src/components/MapView/style.scss
@@ -1,9 +1,1 @@
 @import 'shared/style';
-
-.loft-taxi-map {
-}
-
-.loft-taxi-map-container {
-  width: 100%;
-  height: 100vh - $header-height;
-}
